+++ src/SessionInterface.php
@@ -68,6 +68,11 @@ interface SessionInterface
     public function regenerateId(): void;
 
     /**
+     * Create session ID.
+     */
+    public function createId(): string;
+
+    /**
      * Discard session changes and close session.
      */
     public function discard(): void;