+++ tests/Flash/MockArraySessionStorage.php
@@ -58,6 +58,11 @@ final class MockArraySessionStorage implements SessionInterface
         $this->id = $this->generateId();
     }
 
+    public function createId(): string
+    {
+        return $this->generateId();
+    }
+
     public function discard(): void
     {
         $this->close();